[ADD] #47 - change MapComponent with react-map-gl
