+++ happyfarmer/package.json
@@ -10,12 +10,12 @@
     "@testing-library/user-event": "^13.5.0",
     "apexcharts": "^3.35.3",
     "axios": "^0.27.2",
-    "mapbox-gl": "^2.9.2",
     "rc-slider": "^8.6.6",
     "rc-tooltip": "^3.7.3",
     "react": "^18.2.0",
     "react-apexcharts": "^1.4.0",
     "react-dom": "^18.2.0",
+    "react-map-gl": "^7.0.17",
     "react-redux": "^8.0.2",
     "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
