+++ happyfarmer/src/components/MapComponent.jsx
@@ -1,51 +1,17 @@
 import React, { useRef, useEffect } from 'react';
-
-import mapboxgl from '!mapbox-gl'; // eslint-disable-line import/no-webpack-loader-syntax
-mapboxgl.accessToken = process.env.REACT_APP_MAPBOX_KEY;
+import Map from 'react-map-gl';
 
 export default function MapComponent({ lng, lat, zoom }) {
-  const mapContainer = useRef(null);
-  const map = useRef(null);
-
-  const setMap = (lng, lat) => {
-    if (map.current) return; // initialize map only once
-    map.current = new mapboxgl.Map({
-      container: mapContainer.current,
-      style: 'mapbox://styles/mapbox/streets-v11',
-      center: [lng, lat],
-      zoom: zoom,
-    });
-  };
-
-  const handleGeoSuccess = (pos) => {
-    const lng = pos.coords.longitude;
-    const lat = pos.coords.latitude;
-    setMap(lng, lat);
-  };
-
-  const handleGeoError = (err) => {
-    console.log(err);
-  };
-
-  useEffect(() => {
-    const getGeoLoc = () => {
-      if (!navigator.geolocation) {
-        alert('Geolocation is not supported by your browser');
-      } else {
-        navigator.geolocation.getCurrentPosition(
-          handleGeoSuccess,
-          handleGeoError
-        );
-      }
-    };
-    getGeoLoc();
-  }, []);
-
   return (
-    <div
-      ref={mapContainer}
-      className='map-container'
+    <Map
+      initialViewState={{
+        longitude: lng,
+        latitude: lat,
+        zoom: zoom,
+      }}
       style={{ height: '500px' }}
+      mapStyle='mapbox://styles/mapbox/streets-v9'
+      mapboxAccessToken={process.env.REACT_APP_MAPBOX_KEY}
     />
   );
 }
